requirements.txt
pyrebase4
setuptools
web.py==0.61